requirement.txt
# LangChain Core
 langchain
 langchain-core
 
# OpenAI Integration
 langchain-openai
 openai
 
# Anthropic Integration
 langchain-anthropic
 
# Google Gemini (PaLM) Integrationpi
 langchain-google-genai
 google-generativeai
 
# Environment Variable Management
 python-dotenv
 
# Machine Learning Utilities
 numpy
 scikit-learn